move logger to its own dir
