+++ awebox/mdl/aero/actuator_disk_dir/actuator.py
@@ -33,7 +33,7 @@ _python-3.5 / casadi-3.4.5
 
 import casadi as cas
 import numpy as np
-from awebox.logger import Logger as awelogger
+from awebox.logger.logger import Logger as awelogger
 
 from . import geom as geom
 from . import flow as flow
