+++ awebox/mdl/aero/actuator_disk_dir/geometry_dir/nhat_opt.py
@@ -33,7 +33,7 @@ _python-3.5 / casadi-3.4.5
 
 import numpy as np
 import casadi as cas
-from awebox.logger import Logger as awelogger
+from awebox.logger.logger import Logger as awelogger
 
 import awebox.tools.vector_operations as vect_op
