+++ awebox/ocp/collocation.py
@@ -31,7 +31,7 @@ python-3.5 / casadi-3.4.5
 '''
 
 import casadi.tools as cas
-from awebox.logger import Logger as awelogger
+from awebox.logger.logger import Logger as awelogger
 import numpy as np
 import awebox.tools.struct_operations as struct_op
 from collections import OrderedDict
