+++ awebox/tools/print_operations.py
@@ -28,7 +28,7 @@ _python-3.5 / casadi-3.4.5
 - author:  jochem de schutter 2018
 '''
 
-from awebox.logger import Logger as awelogger
+from awebox.logger.logger import Logger as awelogger
 import os
 
 def print_single_timing(timing):
